# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and inquiry API for Ashley Amani Wambura's professional chef portfolio
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  name: string;
  title: string;
  statement: string;
  phone: string;
  email: string;
  location?: string;
  cvUrl: string;
}

export interface Experience {
  id: number;
  company: string;
  position: string;
  dates: string;
  responsibilities: string[];
}

export interface Education {
  id: number;
  institution: string;
  period: string;
  qualification: string;
  /** @nullable */
  field?: string | null;
}

export interface SkillCategory {
  id: number;
  name: string;
  skills: string[];
}

export interface GalleryItem {
  id: number;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  alt: string;
  featured: boolean;
}

export interface Language {
  id: number;
  language: string;
  proficiency: string;
}

export interface Interest {
  id: number;
  name: string;
}

export interface Reference {
  id: number;
  name: string;
  position: string;
  /** @nullable */
  company?: string | null;
  phone: string;
  email: string;
}

export interface PortfolioSummary {
  experienceCount: number;
  educationCount: number;
  galleryCount: number;
  skillCategoryCount: number;
}

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 160 */
  email: string;
  /** @maxLength 160 */
  company?: string;
  /** @maxLength 160 */
  opportunity?: string;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  message: string;
}

export interface ContactReceipt {
  success: boolean;
  message: string;
}

export type ValidationErrorErrors = {[key: string]: string[]};

export interface ValidationError {
  message: string;
  errors?: ValidationErrorErrors;
}

export type GetGalleryParams = {
category?: string;
};