# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Public content and inquiry API for Ashley Amani Wambura's professional chef portfolio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gallery:
    get:
      operationId: getGallery
      tags: [portfolio]
      summary: List visible gallery items
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit an employment or collaboration inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Inquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactReceipt"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
components:
  schemas:
...
    GalleryItem:
      type: object
      required: [id, title, category, imageUrl, alt, featured]
      properties:
        id: { type: integer }
        title: { type: string }
        category: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: string }
        alt: { type: string }
        featured: { type: boolean }
...
    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 160 }
        company: { type: string, maxLength: 160 }
        opportunity: { type: string, maxLength: 160 }
        message: { type: string, minLength: 10, maxLength: 3000 }
    ContactReceipt:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean }
        message: { type: string }
    ValidationError:
      type: object
      required: [message]
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }