# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and inquiry API for Ashley Amani Wambura's professional chef portfolio
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 160 */
  email: string;
  /** @maxLength 160 */
  company?: string;
  /** @maxLength 160 */
  opportunity?: string;
  /**
     * @minLength 10
     * @maxLength 3000
     */
  message: string;
}