# lib/api-zod/src/generated/types/contactReceipt.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and inquiry API for Ashley Amani Wambura's professional chef portfolio
 * OpenAPI spec version: 0.1.0
 */

export interface ContactReceipt {
  success: boolean;
  message: string;
}