# lib/api-zod/src/generated/types/galleryItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and inquiry API for Ashley Amani Wambura's professional chef portfolio
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItem {
  id: number;
  title: string;
  category: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  alt: string;
  featured: boolean;
}