# lib/api-zod/src/generated/types/validationError.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and inquiry API for Ashley Amani Wambura's professional chef portfolio
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationErrorErrors } from './validationErrorErrors';

export interface ValidationError {
  message: string;
  errors?: ValidationErrorErrors;
}